add metrics data result get api
